Link Point with MeasurementType, and only display available types. gh-4
Add a ManytoMany field to the Point model, linking it to the MeasurementType
model (through Measurement). With corresponding changes in api this allows us
to get a list of which measurement types are present at each point.

Then change map_.html so that only measurement types that are available at the
selected deployment (though not necessarily at all of the points there) are
displayed in the list.