Deal with missing data values. Fixes gh-3
Change the ingestion routine so it doesn't crash when there are missing values
(ie empty cells in the spreadsheet).

It was falling over when trying to save empty measurements, as they are converted
to floats and python can't do float(''). So add a try/exception and don't do
anything with that measurement if it fails.

I also had a test file that was missing some land_dist and theta_is values.
Add a test for these, and replace with fill value -999 if they are missing.