Prevent duplicate entries being created. gh-11
Changes to ingestion routine to prevent it from creating duplicate Campaign,
Deployment, and Point entries.

This is done by creating new subroutines get_campaign etc which do a
Model.objects.get_or_create() using all of the fields to see if there is
already a matching entry.

Note that currently, the MEASUREMENTS will still be duplicated if the same file
is read more than once so some work is still needed.

In addition, new tests are added to check that the duplicates are correctly avoided.
And also added a test for the units_and_name subroutine that was added in
commit bfe590508dcf2ae8dfabcd1d1b015106386e6156