Add initial routines for reading HDF file gh-13
Add first data reading routine, for HDF files. Requires the pyhdf module

Add method for extracting the region of interest. Note that it returns
slice objects rather than lists of points, and that the slice is made from
the min and max i and j indices that were found to be inside the ROI. This is
needed because we don't have a square grid, so a simple list wouldn't be able
to give us a 2d subset array (we would get a 1d list)

Add a high level method ingest_images that would be the one users call, and
it then calls the others as required.