Add regridding option, and method for adding image to database
Add new GeoTools method to regrid data to a regular grid, at the same time as extracting the region of
interest. This simplifies creating the geotiff, as it assumes that all pixels are same size.

Also add the image into the database, as and Image object, with an associated ImageRegion object.