Improve regridding method
Change the way we calculate the grid size to regrid to. Previously we just
used the lon/lat difference between adjacent points on the old grid. This
worked well for the VIIRS test file, but then left holes in AATSR and MERIS
when the data were regridded.

So instead of using adjacent points, we use the lon/lat differences between
opposite corners of squares that are formed by each group of 4 points:

1A------2A------3A
|       |       |
1B------2B------3B
|       |       |
1C------2C------3C

Eg take lon difference between 1A-2B, 2A-3B etc, and
   take lat difference between 2A-1B, 3A-2B etc

As before, we use the maximum value across the grid.