Read in viewing angles + separate file readers for each instrument
Add ability to read in viewing angles, save them to a separate
geotiff, and save the mean value in the database.

As different instruments use different dataset names for the
viewing angles. it was awkward to use the existing file readers.
So, instead of calling a reader for the file type, we call new
readers for individual instruments (which then call the old methods)

Also, as AATSR has separate data arrays for forward and nadir
directions, add some changes to deal with this. We save all the
arrays to the same file, but create two database entries - one
for each direction

Do some refactoring: make metadata, inputdir etc class variables
rather than passing them through as parameters all the time. Now
we pass the IngestImages object instance through to the other classes
instead.