Fix regridding issue that caused empty boxes on the edges
We were getting a row of nans along the  final column and row of the
regridded array. This is because we were only searching within the
min/max values of the coordinates, but by then using int() to calculate
the grid index, we are treating the coord values as the edges of the pixel,
not the centre. Thus we need to search up to "max lon/lat + pixel width"
to fill the end boxes.

The other option was to treat our new lon/lat values as the centre of the
pixel, but GeoTiff specification assumes the corner value it is given is
for the pixel edge, so we'll stick with that.

NB in any case, we are still assuming that the "old" coordinates arrays from
the original input files are for the pixel centres (ie treat them as point
measurements).