Add scaling to HDF files
Read in the scale factor and offset values from HDF file variables, then
apply to data arrays before doing any other processing.

This is not required for the N1 file reader, as there we use the high level
get_band method, instead of get_dataset, and the scaling is done automatically
behind the scenes.