Use flag arrays to mask missing MERIS and AATSR data
For MERIS and AATSR files, read in the flag values and set invalid
pixels (eg padding at the edges) to np.nan. This prevents the regridding
routine using the values on those pixels for interpolation.

Hard code default values for the name of the flag array, and the bit
to use (different bits are different flags - land/ocean, cloud etc), but
these can be overridden by specifiying different ones in the json file.