Tidy up ingested files
After image ingestion, move the data file and metadata file to
a temporary "ingested" folder. (Idea being we can sweep this
directory offline and delete anything over a certain age)
A check is done to see if the geotiff file exists before
moving the files.

The tidyup method needs to know the name of the metadata file,
so add that as an instance attribute.

Also add unit test for the new method (using mock so it doesn't
actually move anything)