More flexible file readers
Add checks to the file readers, such that default values are
used for wavelengths, variable names etc only if they are
not present in the metadata file. So allows users to specify
different settings if required. Default position is to read
reflectance for AATSR & VIIRS, radiance for MERIS (doesn't
have reflectance).