Refactor regridding method to speed it up
Regridding method was very slow, leading to the unit tests taking
a long time.

The problem was the check for whether a point is inside the region
of interest which was done inside the for loops, ie it is evalulated
separately at each grid point. Move outside the for loops and generate
a logical array instead, so it is only done once.

Unit tests now run ~10x quicker.