Fix bug in tidy up
Tidy up method was broken due to the metadata filename variable
containing the path as well as filename, so when we tried to join
the new "ingested" path it created a non existant path. Use
os.path.basemap to strip out the filename before generating
the new path.

(Unit test didn't pick this up as it did the same thing)