Initial changes to documentation, ready for using gh-pages
Remove some hardwired paths, and update author copyright

Also add a shell script to commit changes to documentation to the
gh-pages website. Make your changes on the master branch, then when
ready run the new script to commit them to the gh-pages branch.
nb script only commits, doesn't push up to github. Do that manually
so that you are sure it is ready.