Allow multiple images per JSON file
Change to allow many image files to be described by
one metadata file. read_meta_file now returns a list
of dictionaries, which we loop through.

Format of json files is now:
[
{ image file 1 key:value pairs },
{ image file 2 key:value pairs },
{ etc }
]

This means that even if only one image file is
specified in the json file, it still needs to be
nested within the square brackets, so that the result
is still a list (with one element).