Intermediate work on BRDF tool
Refactor run method so it just takes a json format
object as returned from database query.

Add a main() method to test the tool, while we don't
have the web interface ready.

BRDF tool now extracts reflectance arrays from
the returned list of files, and plots a timeseries
with one line per band.

Add new method to libquerydb to perform a query for
the instrument wavelengths, given an instrument to
search for. We use the returned wavelengths to
label the BRDF plot.

libbrdfroujean now