Return mean dates from the binned timeseries
Instead of storing the mid point of the bin,
calculate the mean date of the included images.
Do this by converting to unix timestamp, so
they are just integers which we can average.

This gave an error when the bin size was such
that we had some bins with no images found,
leading to an empty timestamps list. So add a
check that we found some images before doing
any of the brdf calculations.