Write BRDF output to csv file
Add method to write the BRDF output arrays
to a csv file. File name includes the sensor
name, region and the date range.

The arrays are written with dates along the
columns, and with all bands for first variable
as rows, then all bands for next variable etc
(following DIMITRI output format)