Separate the AATSR directions
We previously had separate database entries for
the two AATSR directions (forward looking and
nadir looking). This meant the geotiff had
2 x nbands arrays in it, but with no way of
labelling which order they come in.

Change so that we create separate geotiff files
(and quicklooks) for each direction, so that
the files are then consistent with other sensors
and have the expected number of arrays inside.

This also means we now use the correct array
of flag values, rather than the using nadir
flags for both directions.