Extract doublets
Add function to libtools to extract doublet pairs
(pairs that match AMC, time, and ROI coverage
criteria).

The function loops through all images from the
target sensor, and finds any reference images
that make a valid doublet with that target image.

Returns lists of the indices of the image pairs in
the original arrays, and a list of the mean time
of each pair.