Radiometric drift for each band
Change radiometric drift routine so that it now
loops over all the bands for the target sensor,
and looks to see if there is a matching band for
the reference sensor (bands are considered matching
if within 10mn). Then calculate the drift for each
band where a match was found.