Fix bug in AATSR image ingestion
A bug in the image ingestion code mean that
both AATSR directions were being assigned the
same viewing angle values, because the variable
names weren't being reset after the first go.

Side effect of this was that we needed a very
large AMC threshold value in radiometric drift
tool in order to find any doublets.

Fix by removing the related keys from the
metadata dictionary before moving on to the
next direction, forcing it to pick up the new
values.

Now we can use amc_threshold=15, as in DIMITRI,
and it correctly finds the doublets.