Refactor calculation of sensor drift
Move the method that calculated drift timeseries
from the Radiometric drift module to libtools, so
we can use it in the supersensor module too. Also
move the associated unit test.

At the same time, change so that instead of just
reference/target, it calculates (target-reference)/reference
and simplify to use a list comprehension instead of
a loop with several lines.