Combine recalibrated data
After looping through all the target sensors,
and doing the recalibration, combine the
data such that each sensor now has the same
number of bands (any bands that weren't present
for that sensor are filled with None).

This is because we can't include band info
in geotiff files, and the database models only
take a single set of wavlengths per sensor. So
we want all the supersensor datasets to cover
the same bands, regardless of which original
sensor they were from.

Add a new method to do the combining, and also
add a unit test for it.